cls_user: always set bucket_count and write header on delete
on removing bucket we need to do a write header as well as otherwise the
bucket count change doesn't get updated. Also since we decrement based
on bucket_count and bucket_count isnt saved in the entry we set this.
Right now we always do a forceful dec_header_stats as we retrieve the
entry from the k-v store where it isn't always set. Need to see better
ways to do this

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>