rgw: implement get/put object tags for S3
This is similar to AWS S3's get/put object tags, tags are stored as an
xattr under the user.rgw.tags key, and are restricted to max 10 tags per
object and key size of 128 and a val size of 256. The API does not reuse
the existing `x-amz-meta` tags. A TODO still to tackle in the API is to
reraise the error 400 when tags are full or exceeding sizes with the
appropriate error messages.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>