rgw: have a configurable authentication order
This implements a configurable authentication order, currently used only for s3
authentication and only supporting external & local authentication, though there
is potential for more finegrained control by allowing for a map of various
engines and the control strategy (required vs sufficient vs fallback)

The current implementation just focuses on setting control sufficient if the
engine is the last in the order (and hence the stack) and just sets fallback to
every other element.

The configuration option is rgw_auth_order which takes a comma/space seperated
list of authentication engines.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>