qa/tasks/{salt,deepsea}: use remote.sh only when parsing output
The remote.sh command does not produce any output in the log until the
command completes. When not parsing the output, it's more important to
get the output logged on a per-line basis: if the command times out,
no output is logged at all, making it extremely difficult to debug the
timeout.

Signed-off-by: Nathan Cutler <ncutler@suse.com>